requirements.txt
alembic==1.4.2
apache-airflow==1.10.11
apispec==1.3.3
argcomplete==1.12.0
asn1crypto==1.4.0
attrs==19.3.0
azure-common==1.1.25
azure-core==1.7.0
azure-storage-blob==12.3.2
Babel==2.8.0
bcrypt==3.1.7
beautifulsoup4==4.9.1
BigQuery-Python==1.15.0
boto3==1.14.33
botocore==1.17.33
cached-property==1.5.1
cachetools==4.1.1
cattrs==1.0.0
certifi==2020.6.20
cffi==1.14.1
chardet==3.0.4
click==7.1.2
colorama==0.4.3
colorlog==4.0.2
configparser==3.5.3
croniter==0.3.34
cryptography==2.9.2
defusedxml==0.6.0
dill==0.3.2
dnspython==2.0.0
docutils==0.15.2
email-validator==1.1.1
Flask==1.1.2
Flask-Admin==1.5.4
Flask-AppBuilder==2.3.4
Flask-Babel==1.0.0
Flask-Caching==1.3.3
Flask-JWT-Extended==3.24.1
Flask-Login==0.4.1
Flask-OpenID==1.2.5
Flask-SQLAlchemy==2.4.4
flask-swagger==0.2.13
Flask-WTF==0.14.3
funcsigs==1.0.2
future==0.18.2
google==3.0.0
google-api-core==1.22.0
google-api-python-client==1.10.0
google-auth==1.20.0
google-auth-httplib2==0.0.4
google-auth-oauthlib==0.4.1
google-cloud==0.34.0
google-cloud-bigquery==1.26.1
google-cloud-bigtable==1.4.0
google-cloud-container==2.0.1
google-cloud-core==1.3.0
google-cloud-dlp==1.0.0
google-cloud-language==1.3.0
google-cloud-secret-manager==1.0.0
google-cloud-spanner==1.17.1
google-cloud-speech==1.3.2
google-cloud-storage==1.30.0
google-cloud-texttospeech==1.0.1
google-cloud-translate==2.0.1
google-cloud-videointelligence==1.15.0
google-cloud-vision==1.0.0
google-crc32c==0.1.0
google-resumable-media==0.7.0
googleapis-common-protos==1.52.0
graphviz==0.14.1
grpc-google-iam-v1==0.12.3
grpcio==1.30.0
grpcio-gcp==0.2.2
gunicorn==19.10.0
httplib2==0.18.1
idna==2.9
iso8601==0.1.12
isodate==0.6.0
itsdangerous==1.1.0
Jinja2==2.10.3
jmespath==0.10.0
json-merge-patch==0.2
jsonschema==3.2.0
lazy-object-proxy==1.5.1
libcst==0.3.8
lockfile==0.12.2
Mako==1.1.3
Markdown==2.6.11
MarkupSafe==1.1.1
marshmallow==2.21.0
marshmallow-enum==1.5.1
marshmallow-sqlalchemy==0.23.1
msrest==0.6.18
mypy-extensions==0.4.3
natsort==7.0.1
numpy==1.19.1
oauthlib==3.1.0
oscrypto==1.2.1
pandas==1.1.0
pandas-gbq==0.13.2
paramiko==2.7.1
pendulum==1.4.4
prison==0.1.3
proto-plus==1.5.1
protobuf==3.12.4
psutil==5.7.2
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
pycryptodomex==3.9.8
pydata-google-auth==1.1.0
Pygments==2.6.1
PyJWT==1.7.1
PyNaCl==1.4.0
pyOpenSSL==19.1.0
pyrsistent==0.16.0
pysnowflake==0.1.3
python-daemon==2.2.4
python-dateutil==2.8.1
python-editor==1.0.4
python-nvd3==0.15.0
python-slugify==4.0.1
python3-openid==3.2.0
pytz==2020.1
pytzdata==2020.1
PyYAML==5.3.1
requests==2.23.0
requests-oauthlib==1.3.0
rsa==4.6
s3transfer==0.3.3
setproctitle==1.1.10
six==1.15.0
snowflake-connector-python==2.2.9
soupsieve==2.0.1
SQLAlchemy==1.3.18
SQLAlchemy-JSONField==0.9.0
SQLAlchemy-Utils==0.36.8
storage==0.0.4.3
tabulate==0.8.7
tenacity==4.12.0
text-unidecode==1.3
thrift==0.13.0
tornado==6.0.4
typing-extensions==3.7.4.2
typing-inspect==0.6.0
tzlocal==1.5.1
unicodecsv==0.14.1
uritemplate==3.0.1
urllib3==1.25.10
Werkzeug==0.16.1
WTForms==2.3.3
zope.deprecation==4.4.0
